Added 'Display Name' feature`
Added the ability to specify a display name in the 'From' header. When specifying msg['To'] it is setting both the "envelope-from" and "From" to the value specified (surrounded by brackets). Added logic to replace the "From" header with Display Name + , if a display name is specified when script is run.